JSON AST parser [WIP]
