+++ runtime/common/src/main/kotlin/kotlinx/serialization/json/JSON.kt
@@ -17,7 +17,6 @@
 package kotlinx.serialization.json
 
 import kotlinx.serialization.*
-import kotlinx.serialization.internal.createString
 import kotlin.reflect.KClass
 
 data class JSON(
@@ -224,7 +223,7 @@ data class JSON(
         override fun readBegin(desc: KSerialClassDesc, vararg typeParams: KSerializer<*>): KInput {
             val newMode = switchMode(mode, desc, typeParams)
             if (newMode.begin != INVALID) {
-                require(p.tc == newMode.beginTc, p.tokenPos) { "Expected '${newMode.begin}, kind: ${desc.kind}'" }
+                p.requireTc(newMode.beginTc) { "Expected '${newMode.begin}, kind: ${desc.kind}'" }
                 p.nextToken()
             }
             return when (newMode) {
@@ -236,7 +235,7 @@ data class JSON(
 
         override fun readEnd(desc: KSerialClassDesc) {
             if (mode.end != INVALID) {
-                require(p.tc == mode.endTc, p.tokenPos) { "Expected '${mode.end}'" }
+                p.requireTc(mode.endTc) { "Expected '${mode.end}'" }
                 p.nextToken()
             }
         }
@@ -246,7 +245,7 @@ data class JSON(
         }
 
         override fun readNullValue(): Nothing? {
-            require(p.tc == TC_NULL, p.tokenPos) { "Expected 'null' literal" }
+            p.requireTc(TC_NULL) { "Expected 'null' literal" }
             p.nextToken()
             return null
         }
@@ -272,7 +271,7 @@ data class JSON(
                         return when (entryIndex++) {
                             0 -> 0
                             1 -> {
-                                require(p.tc == TC_COLON, p.tokenPos) { "Expected ':'" }
+                                p.requireTc(TC_COLON) { "Expected ':'" }
                                 p.nextToken()
                                 1
                             }
@@ -285,7 +284,7 @@ data class JSON(
                     else -> {
                         if (!p.canBeginValue) return READ_DONE
                         val key = p.takeStr()
-                        require(p.tc == TC_COLON, p.tokenPos) { "Expected ':'" }
+                        p.requireTc(TC_COLON) { "Expected ':'" }
                         p.nextToken()
                         val ind = desc.getElementIndex(key)
                         if (ind != UNKNOWN_NAME)
@@ -311,176 +310,6 @@ data class JSON(
 
         override fun <T : Enum<T>> readEnumValue(enumClass: KClass<T>): T = enumFromName(enumClass, p.takeStr())
     }
-
-    private class Parser(val source: String) {
-        var curPos: Int = 0 // position in source
-
-        // updated by nextToken
-        var tokenPos: Int = 0
-        var tc: Byte = TC_EOF
-
-        // update by nextString/nextLiteral
-        var offset = -1 // when offset >= 0 string is in source, otherwise in buf
-        var length = 0 // length of string
-        var buf = CharArray(16) // only used for strings with escapes
-
-        init {
-            nextToken()
-        }
-
-        val canBeginValue: Boolean get() = when (tc) {
-            TC_BEGIN_LIST, TC_BEGIN_OBJ, TC_OTHER, TC_STRING, TC_NULL -> true
-            else -> false
-        }
-
-        fun takeStr(): String {
-            if (tc != TC_OTHER && tc != TC_STRING) fail(tokenPos, "Expected string or non-null literal")
-            val prevStr = if (offset < 0)
-                buf.createString(length) else
-                source.substring(offset, offset + length)
-            nextToken()
-            return prevStr
-        }
-
-        private fun append(ch: Char) {
-            if (length >= buf.size) buf = buf.copyOf(2 * buf.size)
-            buf[length++] = ch
-        }
-
-        // initializes buf usage upon the first encountered escaped char
-        private fun appendRange(source: String, fromIndex: Int, toIndex: Int) {
-            val addLen = toIndex - fromIndex
-            val oldLen = length
-            val newLen = oldLen + addLen
-            if (newLen > buf.size) buf = buf.copyOf(newLen.coerceAtLeast(2 * buf.size))
-            for (i in 0 until addLen) buf[oldLen + i] = source[fromIndex + i]
-            length += addLen
-        }
-
-        fun nextToken() {
-            val source = source
-            var curPos = curPos
-            val maxLen = source.length
-            while(true) {
-                if (curPos >= maxLen) {
-                    tokenPos = curPos
-                    tc = TC_EOF
-                    return
-                }
-                val ch = source[curPos]
-                val tc = c2tc(ch)
-                when (tc) {
-                    TC_WS -> curPos++ // skip whitespace
-                    TC_OTHER -> {
-                        nextLiteral(source, curPos)
-                        return
-                    }
-                    TC_STRING -> {
-                        nextString(source, curPos)
-                        return
-                    }
-                    else -> {
-                        this.tokenPos = curPos
-                        this.tc = tc
-                        this.curPos = curPos + 1
-                        return
-                    }
-                }
-            }
-        }
-
-        private fun nextLiteral(source: String, startPos: Int) {
-            tokenPos = startPos
-            offset = startPos
-            var curPos = startPos
-            val maxLen = source.length
-            while(true) {
-                curPos++
-                if (curPos >= maxLen || c2tc(source[curPos]) != TC_OTHER) break
-            }
-            this.curPos = curPos
-            length = curPos - offset
-            tc = if (rangeEquals(source, offset, length, NULL)) TC_NULL else TC_OTHER
-        }
-
-        private fun nextString(source: String, startPos: Int) {
-            tokenPos = startPos
-            length = 0 // in buffer
-            var curPos = startPos + 1
-            var lastPos = curPos
-            val maxLen = source.length
-            parse@ while (true) {
-                if (curPos >= maxLen) fail(curPos, "Unexpected end in string")
-                if (source[curPos] == STRING) {
-                    break@parse
-                }
-                else if (source[curPos] == STRING_ESC) {
-                    appendRange(source, lastPos, curPos)
-                    val newPos = appendEsc(source, curPos + 1)
-                    curPos = newPos
-                    lastPos = newPos
-                } else {
-                    curPos++
-                }
-            }
-            if (lastPos == startPos + 1) {
-                // there was no escaped chars
-                this.offset = lastPos
-                this.length = curPos - lastPos
-            } else {
-                // some escaped chars were there
-                appendRange(source, lastPos, curPos)
-                this.offset = -1
-            }
-            this.curPos = curPos + 1
-            tc = TC_STRING
-        }
-
-        private fun appendEsc(source: String, startPos: Int): Int {
-            var curPos = startPos
-            require(curPos < source.length, curPos) { "Unexpected end after escape char" }
-            val curChar = source[curPos++]
-            if (curChar == UNICODE_ESC) {
-                curPos = appendHex(source, curPos)
-            } else {
-                val c = esc2c(curChar.toInt())
-                require(c != INVALID, curPos) { "Invalid escaped char '$curChar'" }
-                append(c)
-            }
-            return curPos
-        }
-
-        private fun appendHex(source: String, startPos: Int): Int {
-            var curPos = startPos
-            append(((fromHexChar(source, curPos++) shl 12) +
-                (fromHexChar(source, curPos++) shl 8) +
-                (fromHexChar(source, curPos++) shl 4) +
-                fromHexChar(source, curPos++)).toChar())
-            return curPos
-        }
-
-        fun skipElement() {
-            if (tc != TC_BEGIN_OBJ && tc != TC_BEGIN_LIST) {
-                nextToken()
-                return
-            }
-            val tokenStack = mutableListOf<Byte>()
-            do {
-                when (tc) {
-                    TC_BEGIN_LIST, TC_BEGIN_OBJ -> tokenStack.add(tc)
-                    TC_END_LIST -> {
-                        if (tokenStack.last() != TC_BEGIN_LIST) throw SerializationException("Invalid JSON at $curPos: found ] instead of }")
-                        tokenStack.removeAt(tokenStack.size - 1)
-                    }
-                    TC_END_OBJ -> {
-                        if (tokenStack.last() != TC_BEGIN_OBJ) throw SerializationException("Invalid JSON at $curPos: found } instead of ]")
-                        tokenStack.removeAt(tokenStack.size - 1)
-                    }
-                }
-                nextToken()
-            } while (tokenStack.isNotEmpty())
-        }
-    }
 }
 
 // ----------- JSON utilities -----------
@@ -492,8 +321,8 @@ private enum class Mode(val begin: Char, val end: Char) {
     POLY(BEGIN_LIST, END_LIST),
     ENTRY(INVALID, INVALID);
 
-    val beginTc: Byte = c2tc(begin)
-    val endTc: Byte = c2tc(end)
+    val beginTc: Byte = charToTokenClass(begin)
+    val endTc: Byte = charToTokenClass(end)
 }
 
 private fun switchMode(mode: Mode, desc: KSerialClassDesc, typeParams: Array<out KSerializer<*>>): Mode =
@@ -510,136 +339,18 @@ private fun switchMode(mode: Mode, desc: KSerialClassDesc, typeParams: Array<out
         else -> Mode.OBJ
     }
 
-
-private fun require(condition: Boolean, pos: Int, msg: () -> String ) {
-    if (!condition)
-        fail(pos, msg())
-}
-
-private fun fail(pos: Int, msg: String): Nothing {
-    throw IllegalArgumentException("JSON at $pos: $msg")
-}
-
-// special strings
-private const val NULL = "null"
-
-// special chars
-private const val COMMA = ','
-private const val COLON = ':'
-private const val BEGIN_OBJ = '{'
-private const val END_OBJ = '}'
-private const val BEGIN_LIST = '['
-private const val END_LIST = ']'
-private const val STRING = '"'
-private const val STRING_QUOTE = "\""
-private const val STRING_ESC = '\\'
-
-private const val INVALID = 0.toChar()
-private const val UNICODE_ESC = 'u'
-
-// token classes
-private const val TC_OTHER: Byte = 0
-private const val TC_STRING: Byte = 1
-private const val TC_STRING_ESC: Byte = 2
-private const val TC_WS: Byte = 3
-private const val TC_COMMA: Byte = 4
-private const val TC_COLON: Byte = 5
-private const val TC_BEGIN_OBJ: Byte = 6
-private const val TC_END_OBJ: Byte = 7
-private const val TC_BEGIN_LIST: Byte = 8
-private const val TC_END_LIST: Byte = 9
-private const val TC_NULL: Byte = 10
-private const val TC_INVALID: Byte = 11
-private const val TC_EOF: Byte = 12
-
-// mapping from chars to token classes
-private const val CTC_MAX = 0x7e
-
-private val C2TC = ByteArray(CTC_MAX).apply {
-    for (i in 0..0x20)
-        initC2TC(i, TC_INVALID)
-    initC2TC(0x09, TC_WS)
-    initC2TC(0x0a, TC_WS)
-    initC2TC(0x0d, TC_WS)
-    initC2TC(0x20, TC_WS)
-    initC2TC(COMMA, TC_COMMA)
-    initC2TC(COLON, TC_COLON)
-    initC2TC(BEGIN_OBJ, TC_BEGIN_OBJ)
-    initC2TC(END_OBJ, TC_END_OBJ)
-    initC2TC(BEGIN_LIST, TC_BEGIN_LIST)
-    initC2TC(END_LIST, TC_END_LIST)
-    initC2TC(STRING, TC_STRING)
-    initC2TC(STRING_ESC, TC_STRING_ESC)
-}
-
-private fun ByteArray.initC2TC(c : Int, cl: Byte) { this[c] = cl }
-
-private fun ByteArray.initC2TC(c: Char, cl: Byte) {
-    initC2TC(c.toInt(), cl)
-}
-
-private fun c2tc(c: Char) = if (c.toInt() < CTC_MAX) C2TC[c.toInt()] else TC_OTHER
-
 private fun mustBeQuoted(str: String): Boolean {
     if (str == NULL) return true
     for (ch in str) {
-        if (c2tc(ch) != TC_OTHER) return true
+        if (charToTokenClass(ch) != TC_OTHER) return true
     }
     return false
 }
-
-// mapping from chars to their escape chars and back
-private const val C2ESC_MAX = 0x5d
-private const val ESC2C_MAX = 0x75
-
-private val ESC2C = CharArray(ESC2C_MAX)
-
-private val C2ESC = CharArray(C2ESC_MAX).apply {
-    for (i in 0x00..0x1f)
-        initC2ESC(i, UNICODE_ESC)
-    initC2ESC(0x08, 'b')
-    initC2ESC(0x09, 't')
-    initC2ESC(0x0a, 'n')
-    initC2ESC(0x0c, 'f')
-    initC2ESC(0x0d, 'r')
-    initC2ESC('/', '/')
-    initC2ESC(STRING, STRING)
-    initC2ESC(STRING_ESC, STRING_ESC)
-}
-
-private fun CharArray.initC2ESC(c: Int, esc: Char) {
-    this[c] = esc
-    if (esc != UNICODE_ESC) ESC2C[esc.toInt()] = c.toChar()
-}
-
-private fun CharArray.initC2ESC(c: Char, esc: Char) {
-    initC2ESC(c.toInt(), esc)
-}
-
-private fun esc2c(c: Int): Char = if (c < ESC2C_MAX) ESC2C[c].toChar() else INVALID
 
 private fun toHexChar(i: Int) : Char {
     val d = i and 0xf
     return if (d < 10) (d + '0'.toInt()).toChar()
     else (d - 10 + 'a'.toInt()).toChar()
-}
-
-private fun fromHexChar(source: String, curPos: Int): Int {
-    require(curPos < source.length, curPos) { "Unexpected end in unicode escape" }
-    val curChar = source[curPos]
-    return when (curChar) {
-        in '0'..'9' -> curChar.toInt() - '0'.toInt()
-        in 'a'..'f' -> curChar.toInt() - 'a'.toInt() + 10
-        in 'A'..'F' -> curChar.toInt() - 'A'.toInt() + 10
-        else -> throw fail(curPos, "Invalid toHexChar char '$curChar' in unicode escape")
-    }
-}
-
-private fun rangeEquals(source: String, start: Int, length: Int, str: String): Boolean {
-    val n = str.length
-    if (length != n) return false
-    for (i in 0 until n) if (source[start + i] != str[i]) return false
-    return true
 }
 
 /*
